Reorganize file structure
This makes it more clear what code is used in production vs what is used
for testing.
