--- e2e/hash.test.js
@@ -1,1 +1,0 @@
-../test/hash.test.js
