+++ src/main/common.ts
