+++ tests/e2e/hash.test.js
@@ -1,0 +1,1 @@
+../unit/hash.test.js
