+++ tests/unit/usage.ts
@@ -1,5 +1,5 @@
 // File for testing typescript usage
-import { HighwayHash, WasmHighwayHash, IHash, HashCreator } from "..";
+import { HighwayHash, WasmHighwayHash, IHash, HashCreator } from "../..";
 
 const keyData = Uint8Array.from([
   0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
