+++ tests/verifier/.npmrc
